Implement response types and exceptions

+++ app/Controllers/Http/UsersController.ts
@@ -1,32 +1,44 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 
 import User from 'App/Models/User'
+import { UsersResponse, UserResponse, BasicResponse } from 'App/Controllers/Http/types'
 import CreateUserValidator from 'App/Validators/CreateUserValidator'
 import SortValidator from 'App/Validators/SortValidator'
 import UpdateUserValidator from 'App/Validators/UpdateUserValidator'
+import ModelNotFoundException from 'App/Exceptions/ModelNotFoundException'
 
 export default class UsersController {
   public async index({ request, response }: HttpContextContract) {
     const page = request.input('page', 1) ?? 1
     const limit = request.input('limit', 10) ?? 10
-    const validatedData = await request.validate(SortValidator)
 
+    const validatedData = await request.validate(SortValidator)
     const sortBy = validatedData.sort_by || 'id'
     const order = validatedData.order || 'asc'
 
     const users = await User.query().orderBy(sortBy, order).paginate(page, limit)
+    const result = users.toJSON()
 
-    return response.ok({ users: users })
+    return response.ok({
+      code: 200,
+      message: 'Successfully got users',
+      users: result.data,
+      totalPages: Math.ceil(result.meta.total / limit),
+      currentPage: result.meta.current_page as number,
+    } as UsersResponse)
   }
 
-  public async show({ response, params }: HttpContextContract) {
-    const user = await User.find(params.id)
-
-    if (user) {
-      return response.ok({ user: user })
+  public async show({ params: { id }, response }: HttpContextContract) {
+    const user = await User.find(id)
+    if (!user) {
+      throw new ModelNotFoundException('Invalid id')
     }
 
-    return response.badRequest('User not found')
+    return response.ok({
+      code: 200,
+      message: 'Successfully got user',
+      user: user,
+    } as UserResponse)
   }
 
   public async store({ request, response }: HttpContextContract) {
@@ -34,24 +46,42 @@ export default class UsersController {
 
     const user = await User.create(validatedData)
 
-    return response.created({ user: user })
+    return response.created({
+      code: 201,
+      message: 'Successfully created user',
+      user: user,
+    } as UserResponse)
   }
 
-  public async update({ request, response, params: { id } }: HttpContextContract) {
-    const user = await User.findOrFail(id)
+  public async update({ params: { id }, request, response }: HttpContextContract) {
+    const user = await User.find(id)
+    if (!user) {
+      throw new ModelNotFoundException('Invalid id')
+    }
 
     const validatedData = await request.validate(UpdateUserValidator)
+
     user.merge(validatedData)
     await user.save()
 
-    return response.created({ user: user })
+    return response.created({
+      code: 201,
+      message: 'Successfully updated user',
+      user: user,
+    } as UserResponse)
   }
 
   public async destroy({ params: { id }, response }: HttpContextContract) {
-    const user = await User.findOrFail(id)
+    const user = await User.find(id)
+    if (!user) {
+      throw new ModelNotFoundException('Invalid id')
+    }
 
     await user.delete()
 
-    return response.noContent()
+    return response.ok({
+      code: 200,
+      message: 'Successfully deleted user',
+    } as BasicResponse)
   }
 }
